requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
codecov==2.0.15
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==2.4.2
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-graphql-extensions==0.0.6
django-graphql-jwt==0.2.0
django-model-utils==3.1.2
django-modeltranslation==0.13b1
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
fabric==2.4.0
graphene==2.1.3
graphene-django==2.2.0
graphene-file-upload==1.2.2
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.7
invoke==1.2.0
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
paramiko==2.4.2
Pillow==5.4.1
promise==2.2.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
PyNaCl==1.3.0
pytz==2018.7
PyYAML==4.2b1
requests==2.20.1
requests-oauthlib==1.2.0
rope==0.11.0
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sorl-thumbnail==12.5.0
typed-ast==1.1.0
unittest-xml-reporting==2.2.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
